minor refactoring added integration test
