+++ solr/core/src/test-files/solr/collection1/conf/schema11.xml
@@ -502,6 +502,11 @@ valued. -->
   <dynamicField name="*_dpf" type="delimited_payloads_float" indexed="true"  stored="true"/>
   <dynamicField name="*_dpi" type="delimited_payloads_int" indexed="true"  stored="true"/>
   <dynamicField name="*_dps" type="delimited_payloads_string" indexed="true"  stored="true"/>
+
+  <dynamicField name="*_dpff" type="delimited_payloads_float" indexed="true"  stored="true" multiValued="true"/>
+  <dynamicField name="*_dpii" type="delimited_payloads_int" indexed="true"  stored="true" multiValued="true"/>
+  <dynamicField name="*_dpss" type="delimited_payloads_string" indexed="true"  stored="true" multiValued="true"/>
+
   <fieldType name="delimited_payloads_float" class="solr.TextField" positionIncrementGap="100">
     <analyzer>
       <tokenizer class="solr.MockTokenizerFactory"/>
